requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
datasets==4.5.0
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
filelock==3.24.3
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
nltk==3.9.3
numpy==2.4.2
packaging==26.0
pandas==3.0.1
propcache==0.4.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rouge_score==0.1.2
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
